requirements.txt
# VidCleanser 依赖包
# 核心依赖
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
PyYAML>=6.0
tqdm>=4.65.0

# 可选依赖（用于更好的日志和进度显示）
# rich>=13.0.0
# loguru>=0.7.0

# 图像处理（opencv-python 已包含，但可以单独安装）
# Pillow>=10.0.0

# 注意：
# 1. 需要系统安装 FFmpeg 和 FFprobe
# 2. Windows 用户需要将 FFmpeg 添加到 PATH 环境变量
# 3. 确保 IOPaint 服务已启动并运行在指定端口
